pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr Annenkov",email = "sasha05992@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy==2.0.20",
    "alembic",
    "pydantic",
    "pydantic-settings",
    "fastapi",
    "uvicorn",
    "asyncpg",
    "dotenv",
    "pyjwt",
    "passlib[bcrypt]"

]

[tool.poetry]
packages = [{include = "*", from="bookstore"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
